Work around solution to table default 50 export
Fragile workaround implemented to the 50 row export bug. Cache block
size is changed to the total size of the data and then exported. This is
the only known workaround currently. Cache is then reset to default setting
and datasource. Main problem: setTimeouts are used to coordinate timing of
pagesize change, datasource change and cache reset. Callbacks should be
used instead.