Updated end drag event on d3 simulation and removed forceCenter
End drag was the source of the bug. It based the position of
nodes on its fixed x,y values instead of its non-fixed x,y values. This
made debugging a little tricky. ForceCenter fixed a single node into
the center of the forceCenter. Removing the force resolved this issue.