Resolve initial node render bug in graphview
Selected node Row and its neighbors were being rendered instead of
just the selected nodes. This was because the source node was checked
instead of the target node to see if an expansion should occur. The
source node would always be present because that is the node pushed to
the graph. Thus, the check would always pass in the initial render and
its neighbours would be rendered as well.