Passing graph state via URL works with a couple hiccups
node rid data is encoded and saved via URL. Using nodeRIDs, records
are fetched and processed for graph display. However, in order for
graph states with statement records to be rendered correctly statement
records must be processed last. This is because of link properties
which are unidirectional. This has caused coloring issues due to
differing node processing order.