Implement graph state saving and sharing via encoding node data in URL
Due to limited length of URL and in order to maximize number of nodes
that could be shared, graph nodes are the only encoded bits of data.
This caused an inconsistent coloring state of the graph when displaying
nodes with link properties due to its uni-directional nature. To resolve
this bug, nodes with link properties are rendered last so that their
linked records already exist in the graph to ensure correct link display.